Refactor: Organize training code under crfs::train submodule
- Move training-related modules to src/train/ directory
  - dictionary.rs
  - feature_gen.rs
  - crf_context.rs
  - trainer.rs
  - model_writer.rs
- Create train/mod.rs with public API exports
- Update lib.rs to expose train submodule
- Update all imports in tests and examples
- Maintain backward compatibility with re-exports
- All 21 tests passing

This improves code organization by clearly separating training
and prediction functionality into distinct modules.
